use type api.RepoID (int32) for repository database IDs
This makes the naming of parameters and fields clearer and more consistent.

+++ cmd/frontend/internal/graphqlbackend/org_repos.go
@@ -14,7 +14,7 @@ type orgRepoResolver struct {
 }
 
 func (o *orgRepoResolver) ID() int32 {
-	return o.repo.ID
+	return int32(o.repo.ID)
 }
 
 func (o *orgRepoResolver) Org() *orgResolver {
